+++ eladmin-gateway/src/main/java/org/micah/gateway/mapper/RouterMapper.java
@@ -26,5 +26,5 @@ public interface RouterMapper extends BaseMapper<Router> {
      * @param id
      * @return
      */
-    Router getById(Integer id);
+    Router getById(Long id);
 }
